- finished generalization: one map card now is used for all map types
